refactor: 마켓페이지 전체 무한 스크롤  적용 및 필터 카운트 유틸함수 적용
